Refactored and fix piece ascii
refactored the code into multiple files for organization and fixed ascii art for pieces to make them scale dynamically with board size
